requirements.txt
google-adk>=0.4.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
requests>=2.31.0
requests>=2.32.3
